requirements.txt
yfinance
pandas
numpy
scikit-learn
matplotlib
requests
beautifulsoup4
pandas-ta
streamlit
plotly
newsapi-python